# docs/swagger.yaml
basePath: /v1
definitions:
...
  models.DeleteEmployee:
    properties:
      msnv:
        type: string
    required:
    - msnv
    type: object
...
  models.DoorlockStatusLog:
    properties:
      doorId:
        type: string
      id:
        type: integer
      stateValue:
        description: corresponding statetype
        type: string
      statusType:
        description: ConnectState, DoorState, LockState
        type: string
    type: object
  models.Employee:
    properties:
      department:
        type: string
      email:
        type: string
      highestPriority:
        type: boolean
      id:
        type: integer
      keypadPass:
        type: string
      msnv:
        type: string
      name:
        type: string
      phone:
        type: string
      rfidPass:
        type: string
      role:
        type: string
      schedulers:
        items:
          $ref: '#/definitions/models.Scheduler'
        type: array
    required:
    - msnv
    type: object
...
  models.Scheduler:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      customerId:
        type: string
      doorId:
        description: DoorID          *string `gorm:"type:varchar(256);" json:"doorId"`
        type: integer
      employeeId:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      role:
        description: value in ["employee", "student", "customer"]
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      studentId:
        type: string
      userId:
        type: string
      weekDay:
        type: integer
    type: object
...
  models.SwagCreateEmployee:
    properties:
      department:
        type: string
      email:
        type: string
      highestPriority:
        type: boolean
      keypadPass:
        type: string
      msnv:
        type: string
      name:
        type: string
      phone:
        type: string
      rfidPass:
        type: string
      role:
        type: string
    required:
    - msnv
    type: object
...
  models.UserSchedulerReq:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      customerId:
        type: string
      doorId:
        description: DoorID          *string `gorm:"type:varchar(256);" json:"doorId"`
        type: integer
      doorlockAddress:
        type: string
      employeeId:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      gatewayId:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      role:
        description: value in ["employee", "student", "customer"]
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      studentId:
        type: string
      userId:
        type: string
      weekDay:
        type: integer
    required:
    - doorlockAddress
    - gatewayId
    type: object
  utils.ErrorResponse:
    properties:
      errorMsg:
        type: string
      msg:
        type: string
      statusCode:
        type: integer
    type: object
host: http://iot.hcmue.space:8002
info:
  contact:
    name: Mr. Khai
  description: This is API document for DMS backend server
  title: Backend API
  version: "1.0"
paths:
...
  /v1/doorlockStatusLog/:fromTime/:toTime:
    delete:
      description: delete doorlock status logs in time range
      parameters:
      - description: From Unix time
        in: path
        name: fromTime
        required: true
        type: string
      - description: To Unix time
        in: path
        name: toTime
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Delete DoorlockStatusLog In Time Range
    get:
      description: find doorlock status logs in time range
      parameters:
      - description: From Unix time
        in: path
        name: fromTime
        required: true
        type: string
      - description: To Unix time
        in: path
        name: toTime
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DoorlockStatusLog'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find DoorlockStatusLog In Time Range
  /v1/doorlockStatusLog/{doorId}:
    get:
      description: find doorlock status logs by door id
      parameters:
      - description: DoorlockStatusLog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DoorlockStatusLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find DoorlockStatusLog By DoorID
  /v1/doorlockStatusLog/door/:id:
    delete:
      description: delete doorlock status logs by doorId
      parameters:
      - description: DoorID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Delete Doorlock status logs by doorId
  /v1/doorlockStatusLogs:
    get:
      description: find all doorlock status logs info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.DoorlockStatusLog'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find All DoorlockStatusLog
...
  /v1/employee:
    delete:
      consumes:
      - application/json
      description: Delete employee using "msnv" field. Send deleted info to MQTT broker
      parameters:
      - description: Employee MSNV
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.DeleteEmployee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Delete Employee By MSNV
    patch:
      consumes:
      - application/json
      description: Update employee, must have correct "id" and "msnv" field. Send
        updated info to MQTT broker
      parameters:
      - description: Fields need to update an employee
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Update Employee By ID and MSNV
    post:
      consumes:
      - application/json
      description: Create employee. Send created info to MQTT broker
      parameters:
      - description: Fields need to create a employee
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.SwagCreateEmployee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Create Employee
  /v1/employee/{msnv}:
    get:
      description: find employee info by employee msnv
      parameters:
      - description: Employee MSNV
        in: path
        name: msnv
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find Employee By MSNV
  /v1/employee/{msnv}/scheduler:
    post:
      consumes:
      - application/json
      description: Add scheduler that allows employee open specific door. Send updated
        info to MQTT broker
      parameters:
      - description: Request with Scheduler, GatewayID, DoorlockAdress
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UserSchedulerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Add Door Open Scheduler For Employee
  /v1/employees:
    get:
      description: find all employees info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Employee'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find All Employee
...
swagger: "2.0"